kernel: core_hook: provide a better reboot handler (#523)
* Revert "feat: try manual reboot hook (#521)"

This reverts commit 1853d9decf3a684a547f97b81c8551edfad05ca4.

* kernel: core_hook: provide a better reboot handler

I propose that you pass cmd and arg as reference.
this is so we can have much more extendable use of that pointer

kernel: core_hook: provide sys_reboot handler
- https://github.com/backslashxx/KernelSU/commit/2e2727d56c68cd3c0a724bee7b04ec32c4786546

kernel: kp_ksud: add sys_reboot kp hook
- https://github.com/backslashxx/KernelSU/commit/03285886b0822db283f0dff0ffffe3223f5a6988

I'm proposing passing arg as reference to arg pointer and also pass int cmd
we can use it to pass numbers atleast.
for advanced usage, we can use it as a delimiter so we can pass a pointer to array.

example pass a char *array[] which decays to a char ** and then use cmd as the number of array members.
we can pass the pointer of the first member of the array and use cmd as the delimiter (count) of members.

for simpler usecase, heres some that I added.

kernel: core_hook: expose  umount list on sys_reboot interface
- https://github.com/backslashxx/KernelSU/commit/352de41e4beb5950288d27d2f84aa1a5bda9fa59

kernel: core_hook: expose nuke_ext4_sysfs to sys_reboot interface
- https://github.com/backslashxx/KernelSU/commit/83fc684ccbf8b57bb707997c04e554965215d334

ksud: add cmd for add-try-umount, wipe-umount-list and nuke-ext4-sysfs
- https://github.com/backslashxx/KernelSU/commit/a4eab4b8c33cd0893756d9a0e9a6ec675ef868a3

more usage demos
https://github.com/backslashxx/lkm_template/tree/write-pointer-on-pointer
https://github.com/backslashxx/lkm_template/tree/pointer-reuse

I actually proposed sys_reboot upstream because of this pointer that is very usable.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>

---------

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>