Rework sulog (#744)
* Remove sulog

* kernel: sulog: basic ringbuffer, timestamped su log

This is a very basic ringbuffer implementation.
on this one we keep 8 bytes per entry and 250 entries max.

This logs escalation access about: who did it, what happened and when did
it happen.

userspace can deduce as we also send next index to it. which makes it so
that this is also the oldest entry. next -1 is the latest entry.

v2: timestamped version

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>

* manager: Redesign the user interface for sulog

* kernel: fix build & refactor

* Minor optimizations

---------

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>
Co-authored-by: backslashxx <118538522+backslashxx@users.noreply.github.com>